added home card component and reorganized screens
